// apps/web/src/features/chat/types.ts
export type MsgToolCallPart = {
  type: "tool-call"
  toolName: string
  toolCallId: string
  input: unknown
  isStreaming?: boolean
}

export type ToolResult =
  | {
      type: "text"
      value: string
    }
  | {
      type: "error-text"
      value: string
    }

export type MsgToolResultPart = {
  type: "tool-result"
  toolName: string
  toolCallId: string
  output: ToolResult
  isStreaming?: boolean
}